requirements.txt
# Core dependencies
tree-sitter==0.20.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.7.0
pydantic-settings==2.6.1
jinja2==3.1.2
numpy==1.26.2
pandas==2.1.3
mermaid-py==0.1.1
